Convert EncryptRuleConfiguration to DistSQL via SPI
